django-back-end/requirements.txt
django==3.1.5
djangorestframework==3.12.2
django-cors-headers==3.7.0